requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.0.0
mediapipe>=0.10.0
ultralytics>=8.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
websockets>=12.0
